requirements.txt
torch==2.2.2 --index-url https://download.pytorch.org/whl/cu121
torchvision==0.17.2 --index-url https://download.pytorch.org/whl/cu121
xformers==0.0.25.post1 --index-url https://download.pytorch.org/whl/cu121


beartype==0.18.5
einops==0.8.0
einops-exts==0.0.4
opencv-python==4.9.0.80
pillow==10.3.0


colossalai>=0.4.1
mmengine>=0.10.3
pandas>=2.0.3
timm==0.9.16
rotary_embedding_torch==0.5.3
ftfy>=6.2.0 # for t5
diffusers==0.29.0 # for vae
accelerate==0.29.2 # for t5
av>=12.0.0 # for video loading
numpy<2.0.0

# [gradio]
gradio>=4.26.0
spaces>=0.28.3

# [notebook]
ipykernel>=6.29.4
ipywidgets>=8.1.2

# [training]
wandb>=0.17.0
tensorboard>=2.14.0
pandarallel>=1.6.5
pyarrow>=16.1.0 # for parquet

# [dev]
pre-commit>=3.5.0
openai